+++ rotkehlchen/history/events.py
@@ -18,7 +18,7 @@ from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.typing import EthereumTransaction, Location, Timestamp
 from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.accounting import action_get_timestamp
-from rotkehlchen.utils.misc import ts_now
+from rotkehlchen.utils.misc import ts_now, timestamp_to_date
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.manager import ChainManager
@@ -82,11 +82,24 @@ class EventsHistorian():
         self.db = db
         self.exchange_manager = exchange_manager
         self.chain_manager = chain_manager
-        self._reset_progress_variables()
+        db_settings = self.db.get_settings()
+        self.dateformat = db_settings.date_display_format
+        self.datelocaltime = db_settings.display_date_in_localtime
+        self._reset_variables()
 
-    def _reset_progress_variables(self) -> None:
+    def timestamp_to_date(self, timestamp: Timestamp) -> str:
+        return timestamp_to_date(
+            timestamp,
+            formatstr=self.dateformat,
+            treat_as_local=self.datelocaltime,
+        )
+
+    def _reset_variables(self) -> None:
         self.processing_state_name = 'Starting query of historical events'
         self.progress = ZERO
+        db_settings = self.db.get_settings()
+        self.dateformat = db_settings.date_display_format
+        self.datelocaltime = db_settings.display_date_in_localtime
 
     def _increase_progress(self, step: int, total_steps: int) -> int:
         step += 1
@@ -122,7 +135,7 @@ class EventsHistorian():
             has_premium: bool,
     ) -> HistoryResult:
         """Creates trades and loans history from start_ts to end_ts"""
-        self._reset_progress_variables()
+        self._reset_variables()
         step = 0
         total_steps = len(self.exchange_manager.connected_exchanges) + HISTORY_QUERY_STEPS
         log.info(
@@ -229,14 +242,22 @@ class EventsHistorian():
                 from_ts=start_ts,
                 to_ts=end_ts,
             )
-            for gain, timestamp in dsr_gains:
-                if gain > ZERO:
-                    defi_events.append(DefiEvent(
-                        timestamp=timestamp,
-                        event_type=DefiEventType.DSR_LOAN_GAIN,
-                        asset=A_DAI,
-                        amount=gain,
-                    ))
+            for gain in dsr_gains:
+                if gain.amount <= ZERO:
+                    continue
+
+                notes = (
+                    f'MakerDAO DSR Gains from {self.timestamp_to_date(gain.from_timestamp)}'
+                    f' to {self.timestamp_to_date(gain.to_timestamp)}'
+                )
+                defi_events.append(DefiEvent(
+                    timestamp=gain.to_timestamp,
+                    event_type=DefiEventType.DSR_LOAN_GAIN,
+                    asset=A_DAI,
+                    amount=gain.amount,
+                    tx_hashes=gain.tx_hashes,
+                    notes=notes,
+                ))
         step = self._increase_progress(step, total_steps)
 
         # Include makerdao vault events
@@ -250,11 +271,20 @@ class EventsHistorian():
             for detail in vault_details:
                 last_event_ts = detail.events[-1].timestamp
                 if start_ts <= last_event_ts <= end_ts:
+                    notes = (
+                        f'USD value of DAI lost for MakerDAO vault {detail.identifier} '
+                        f'due to accrued debt or liquidations. IMPORTANT: At the moment rotki '
+                        f'can\'t figure debt until a given time, so this is debt until '
+                        f'now. If you are looking at a past range this may be bigger '
+                        f'than it should be. We are actively working on improving this'
+                    )
                     defi_events.append(DefiEvent(
                         timestamp=last_event_ts,
                         event_type=DefiEventType.MAKERDAO_VAULT_LOSS,
                         asset=A_USD,
                         amount=detail.total_liquidated.usd_value + detail.total_interest_owed,
+                        tx_hashes=[x.tx_hash for x in detail.events],
+                        notes=notes,
                     ))
         step = self._increase_progress(step, total_steps)
 
@@ -268,8 +298,8 @@ class EventsHistorian():
                 from_timestamp=start_ts,
                 to_timestamp=end_ts,
             )
-            for _, vault_mappings in yearn_vaults_history.items():
-                for _, vault_history in vault_mappings.items():
+            for address, vault_mappings in yearn_vaults_history.items():
+                for vault_name, vault_history in vault_mappings.items():
                     # For the vaults since we can't get historical values of vault tokens
                     # yet, for the purposes of the tax report count everything as USD
                     for yearn_event in vault_history.events:
@@ -279,6 +309,11 @@ class EventsHistorian():
                                 event_type=DefiEventType.YEARN_VAULTS_PNL,
                                 asset=A_USD,
                                 amount=yearn_event.realized_pnl.usd_value,
+                                tx_hashes=[yearn_event.tx_hash],
+                                notes=(
+                                    f'USD equivalent PnL for {address} and yearn '
+                                    f'{vault_name} at event'
+                                ),
                             ))
         step = self._increase_progress(step, total_steps)
 
@@ -306,6 +341,11 @@ class EventsHistorian():
                         event_type=DefiEventType.COMPOUND_LOAN_INTEREST,
                         asset=event.to_asset,
                         amount=event.realized_pnl.amount,
+                        tx_hashes=[event.tx_hash],
+                        notes=(
+                            f'Interest earned in compound for '
+                            f'{event.to_asset.identifier} until this event'
+                        ),
                     ))
                 elif event.event_type == 'repay':
                     defi_events.append(DefiEvent(
@@ -313,6 +353,11 @@ class EventsHistorian():
                         event_type=DefiEventType.COMPOUND_DEBT_REPAY,
                         asset=event.asset,
                         amount=event.realized_pnl.amount,
+                        tx_hashes=[event.tx_hash],
+                        notes=(
+                            f'Amount of {event.asset.identifier} lost in '
+                            f'compound due to debt repayment'
+                        ),
                     ))
                 elif event.event_type == 'liquidation':
                     defi_events.append(DefiEvent(
@@ -320,12 +365,22 @@ class EventsHistorian():
                         event_type=DefiEventType.COMPOUND_LIQUIDATION_DEBT_REPAID,
                         asset=event.asset,
                         amount=event.value.amount,
+                        tx_hashes=[event.tx_hash],
+                        notes=(
+                            f'Amount of {event.asset.identifier} gained in '
+                            f'compound due to liquidation debt repayment'
+                        ),
                     ))
                     defi_events.append(DefiEvent(
                         timestamp=event.timestamp,
                         event_type=DefiEventType.COMPOUND_LIQUIDATION_COLLATERAL_LOST,
                         asset=event.to_asset,
                         amount=event.to_value.amount,
+                        tx_hashes=[event.tx_hash],
+                        notes=(
+                            f'Amount of {event.to_asset.identifier} collateral lost '
+                            f'in compound due to liquidation'
+                        ),
                     ))
                 elif event.event_type == 'comp':
                     defi_events.append(DefiEvent(
@@ -333,6 +388,7 @@ class EventsHistorian():
                         event_type=DefiEventType.COMPOUND_REWARDS,
                         asset=event.asset,
                         amount=event.realized_pnl.amount,
+                        tx_hashes=[event.tx_hash],
                     ))
         step = self._increase_progress(step, total_steps)
 
@@ -347,18 +403,23 @@ class EventsHistorian():
                 to_timestamp=end_ts,
             )
             for _, adex_history in adx_mapping.items():
+                # The transaction hashes here are not accurate. Need to figure out
+                # a way to have accurate transaction hashes for events in a time period
+                adex_tx_hashes = [x.tx_hash for x in adex_history.events]
                 for adx_detail in adex_history.staking_details:
                     defi_events.append(DefiEvent(
                         timestamp=end_ts,
                         event_type=DefiEventType.ADEX_STAKE_PROFIT,
                         asset=A_ADX,
                         amount=adx_detail.adx_profit_loss.amount,
+                        tx_hashes=adex_tx_hashes,  # type: ignore
                     ))
                     defi_events.append(DefiEvent(
                         timestamp=end_ts,
                         event_type=DefiEventType.ADEX_STAKE_PROFIT,
                         asset=A_DAI,
                         amount=adx_detail.dai_profit_loss.amount,
+                        tx_hashes=adex_tx_hashes,  # type: ignore
                     ))
         step = self._increase_progress(step, total_steps)
 
@@ -389,35 +450,64 @@ class EventsHistorian():
                             event_type=DefiEventType.AAVE_LOAN_INTEREST,
                             asset=event.asset,
                             amount=event.value.amount,
+                            tx_hashes=[event.tx_hash],
                         ))
                         total_amount_per_token[event.asset] += event.value.amount
 
-                for token, balance in aave_history.total_earned_interest.items():
-                    # Αdd an extra event per token per address for the remaining not paid amount
-                    if token in total_amount_per_token:
-                        defi_events.append(DefiEvent(
-                            timestamp=now,
-                            event_type=DefiEventType.AAVE_LOAN_INTEREST,
-                            asset=token,
-                            amount=balance.amount - total_amount_per_token[token],
-                        ))
+                # TODO: Here we should also calculate any unclaimed interest payments
+                # within the time range. IT's quite complicated to do that though and
+                # would most probably require an archive node
 
                 # Add all losses from aave borrowing/liquidations
                 for asset, balance in aave_history.total_lost.items():
+                    aave_tx_hashes = []
+                    for event in aave_history.events:
+                        if event not in ('borrow', 'repay', 'liquidation'):
+                            continue
+
+                        if event in ('borrow', 'repay') and event.asset == asset:
+                            aave_tx_hashes.append(event.tx_hash)
+                            continue
+
+                        relevant_liquidation = (
+                            event.event_type == 'liquidation' and
+                            asset in (event.collateral_asset, event.principal_asset)
+                        )
+                        if relevant_liquidation:
+                            aave_tx_hashes.append(event.tx_hash)
+
                     defi_events.append(DefiEvent(
                         timestamp=now,
                         event_type=DefiEventType.AAVE_LOSS,
                         asset=asset,
                         amount=balance.amount,
+                        tx_hashes=aave_tx_hashes,
+                        notes=(
+                            f'All {asset.identifier} lost in Aave due to borrowing '
+                            f'debt or liquidations in the PnL period.'
+                        ),
                     ))
 
                 # Add earned assets from aave liquidations
                 for asset, balance in aave_history.total_earned_liquidations.items():
+                    aave_tx_hashes = []
+                    for event in aave_history.events:
+                        relevant_liquidation = (
+                            event.event_type == 'liquidation' and
+                            asset in (event.collateral_asset, event.principal_asset)
+                        )
+                        if relevant_liquidation:
+                            aave_tx_hashes.append(event.tx_hash)
+
                     defi_events.append(DefiEvent(
                         timestamp=now,
                         event_type=DefiEventType.AAVE_LOAN_INTEREST,
                         asset=asset,
                         amount=balance.amount,
+                        tx_hashes=aave_tx_hashes,
+                        notes=(
+                            f'All {asset.identifier} gained in Aave due to liquidation leftovers'
+                        ),
                     ))
         self._increase_progress(step, total_steps)
         history.sort(key=action_get_timestamp)